+++ src/main/java/com/lowdragmc/mbd2/integration/create/machine/ConfigKineticMachineSettings.java
@@ -1,5 +1,6 @@
 package com.lowdragmc.mbd2.integration.create.machine;
 
+import com.jozufozu.flywheel.backend.RenderLayer;
 import com.lowdragmc.lowdraglib.gui.editor.annotation.Configurable;
 import com.lowdragmc.lowdraglib.gui.editor.annotation.NumberRange;
 import com.lowdragmc.lowdraglib.gui.editor.configurator.IConfigurable;
@@ -68,6 +69,10 @@ public class ConfigKineticMachineSettings implements IConfigurable, IPersistedSe
     @Configurable(name = "config.kinetic_machine.use_flywheel", tips = "config.kinetic_machine.use_flywheel.tooltip")
     @Builder.Default
     public boolean useFlywheel = true;
+    @Configurable(name = "config.kinetic_machine.flywheel_render_layer", tips = "config.kinetic_machine.flywheel_render_layer.tooltip")
+    @Builder.Default
+    public RenderLayer renderLayer = RenderLayer.SOLID;
+
     /**
      * Get the rotation facing based on the machine front facing;
      */