+++ src/main/java/com/lowdragmc/mbd2/integration/create/machine/MBDKineticInstance.java
@@ -2,6 +2,7 @@ package com.lowdragmc.mbd2.integration.create.machine;
 
 import com.jozufozu.flywheel.api.InstanceData;
 import com.jozufozu.flywheel.api.MaterialManager;
+import com.jozufozu.flywheel.backend.RenderLayer;
 import com.jozufozu.flywheel.core.PartialModel;
 import com.simibubi.create.content.kinetics.base.KineticBlockEntityInstance;
 import com.simibubi.create.content.kinetics.base.flwdata.RotatingData;
@@ -16,12 +17,13 @@ public class MBDKineticInstance extends KineticBlockEntityInstance<MBDKineticMac
 
     protected final List<RotatingData> keys = new ArrayList<>();
 
-    public MBDKineticInstance(MaterialManager modelManager, MBDKineticMachineBlockEntity tile, PartialModel model) {
+    public MBDKineticInstance(MaterialManager modelManager, MBDKineticMachineBlockEntity tile, PartialModel model, RenderLayer layer) {
         super(modelManager, tile);
         var speed = tile.getSpeed();
         var rotationFacing = tile.definition.kineticMachineSettings.getRotationFacing(tile.getMetaMachine().getFrontFacing().orElse(Direction.NORTH));
-        var rotatingData = materialManager
-                .defaultSolid()
+        var rotatingData = (layer == RenderLayer.SOLID ? modelManager.defaultSolid() :
+                        layer == RenderLayer.CUTOUT ? modelManager.defaultCutout() :
+                        layer == RenderLayer.TRANSPARENT ? modelManager.defaultTransparent() : modelManager.defaultSolid())
                 .material(AllMaterialSpecs.ROTATING)
                 .getModel(model, blockState, rotationFacing, CachedBufferer.rotateToFaceVertical(rotationFacing))
                 .createInstance();