+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 * add item durability recipe capability.
 * add create rpm recipe capability.
+* add custom render layer for flywheel instance
 * remove duplicated nodes.